requirements.txt
astor==0.8.1
astunparse==1.6.3
blinker==1.4
brotlipy==0.7.0
cycler==0.11.0
flasgger==0.9.5
Flask==2.0.3
fonttools==4.27.1
imutils==0.5.4
itsdangerous==2.0.1
joblib==1.1.0
JSON-minify==0.3.0
kiwisolver==1.3.1
matplotlib==3.3.4
mkl-fft==1.0.6
scikit-build
nbconvert==5.5.0
opencv-contrib-python==3.4.4.19
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.4
pyasn1-modules==0.2.8
pytz==2022.1
PyYAML==6.0
requests-oauthlib==1.3.0
scikit-learn==0.24.2
sip==6.5.1

termcolor==1.1.0
threadpoolctl==3.1.0
webencodings==0.5.1

